requirements.txt
django==5.0.4
duckdb==0.10.1
jupyter==1.0.0
numpy==1.26.4
pandas==2.2.1
psycopg2-binary==2.9.9
pyarrow==15.0.2
python-dateutil==2.9.0.post0
requests==2.31.0
requests-file==2.0.0
python-environ==0.4.54
pydantic==2.7.0
django-environ==0.11.2
celery[redis]==5.4.0
django-celery-beat==2.6.0
pandera==0.18.3
django-pgtrigger==4.11.1